pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "uncle-joes-coffee-shop-api"
version = "1.0.0"
description = "Production-ready FastAPI backend for Uncle Joe's Coffee Shop."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.116.1",
  "google-cloud-bigquery==3.37.0",
  "pydantic==2.11.7",
  "pydantic-settings==2.10.1",
  "uvicorn[standard]==0.35.0",
]

[project.optional-dependencies]
dev = [
  "httpx==0.28.1",
  "pytest==8.4.1",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
